requirements.txt
fastapi
uvicorn
pydantic-settings
openai
langchain
python-dotenv
ruff
opencv-python
Pillow
numpy
mediapipe